app/backend/pyproject.toml
[project]
name = "knowledge-ddd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.42.86",
    "botocore[crt]>=1.42.86",
    "certifi>=2026.2.25",
    "colorlogger>=1.0",
    "fastapi[standard]>=0.135.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.24",
    "uvicorn>=0.44.0",
]
